Added editor files
Added files needed to start developing a level editor for the game. They're mostly the same thing, but are built specifically for an editor. LevelEditor.html has mostly the same file dependencies as the normal game, but it has variations of some of them such as EditorMain.js, EditorControls.js, and EditorConstructors.js. I have them added as separate files because the editor won't need some of the normal files like initialize and main, but it will have its own version of that stuff. It's going to have different ways to control the game and creation of objects, etc.