Slight refactoring, added new constructors
I took the player movement lines out of its function in player.update because it was weird, I had to pass itself in as a parameter to use it, I didn't like it much. Since it was mainly for organization, I took it all out and settled on using comments instead. As for the classes, I added the skeletons of the new ones, Sector, which is the same as something like a world in other video games; a container for levels. I added level, to contain all of the level stuff, level data, just a more specific thing that will eventually go inside of level, and platform group, which is for the most part a container for platform objects. Additionally, I am now using strict mode.